Refactored IBogeyBlock to AbstractBogeyBlock and extracted relevant StandardBogeyBlock implementations
