+++ src/main/java/com/simibubi/create/content/logistics/trains/track/StandardBogeyBlock.java
@@ -2,93 +2,32 @@ package com.simibubi.create.content.logistics.trains.track;
 
 import java.util.EnumSet;
 
-import com.jozufozu.flywheel.api.MaterialManager;
-import com.mojang.blaze3d.vertex.PoseStack;
-import com.mojang.blaze3d.vertex.VertexConsumer;
-import com.mojang.math.Vector3f;
-import com.simibubi.create.AllBlockPartials;
 import com.simibubi.create.AllBlocks;
+import com.simibubi.create.AllBogeyStyles;
 import com.simibubi.create.AllTileEntities;
-import com.simibubi.create.content.contraptions.relays.elementary.ShaftBlock;
 import com.simibubi.create.content.logistics.trains.BogeyRenderer;
-import com.simibubi.create.content.logistics.trains.IBogeyBlock;
-import com.simibubi.create.content.logistics.trains.StandardBogeyRenderer;
-import com.simibubi.create.content.logistics.trains.entity.BogeyInstance;
-import com.simibubi.create.content.logistics.trains.entity.CarriageBogey;
-import com.simibubi.create.content.logistics.trains.entity.StandardBogeyInstance;
+import com.simibubi.create.content.logistics.trains.AbstractBogeyBlock;
+import com.simibubi.create.content.logistics.trains.entity.BogeyStyle;
 import com.simibubi.create.content.schematics.ISpecialBlockItemRequirement;
-import com.simibubi.create.content.schematics.ItemRequirement;
-import com.simibubi.create.content.schematics.ItemRequirement.ItemUseType;
 import com.simibubi.create.foundation.block.ITE;
 import com.simibubi.create.foundation.block.ProperWaterloggedBlock;
-import com.simibubi.create.foundation.render.CachedBufferer;
-import com.simibubi.create.foundation.utility.AngleHelper;
-import com.simibubi.create.foundation.utility.Iterate;
 
-import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.client.renderer.RenderType;
 import net.minecraft.core.BlockPos;
-import net.minecraft.core.Direction;
-import net.minecraft.core.Direction.Axis;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.BlockGetter;
-import net.minecraft.world.level.LevelAccessor;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.Rotation;
-import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.StateDefinition.Builder;
-import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraft.world.level.block.state.properties.EnumProperty;
-import net.minecraft.world.level.material.FluidState;
 import net.minecraft.world.phys.HitResult;
 import net.minecraft.world.phys.Vec3;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 
-public class StandardBogeyBlock extends Block
-	implements IBogeyBlock, ITE<StandardBogeyTileEntity>, ProperWaterloggedBlock, ISpecialBlockItemRequirement {
-
-	public static final EnumProperty<Axis> AXIS = BlockStateProperties.HORIZONTAL_AXIS;
+public class StandardBogeyBlock extends AbstractBogeyBlock implements ITE<StandardBogeyTileEntity>, ProperWaterloggedBlock, ISpecialBlockItemRequirement {
 	private final boolean large;
-
-	private final BogeyRenderer renderer;
 
 	public StandardBogeyBlock(Properties p_i48440_1_, boolean large) {
 		super(p_i48440_1_);
 		this.large = large;
-		this.renderer = new StandardBogeyRenderer();
 		registerDefaultState(defaultBlockState().setValue(WATERLOGGED, false));
-	}
-
-	@Override
-	protected void createBlockStateDefinition(Builder<Block, BlockState> builder) {
-		builder.add(AXIS, WATERLOGGED);
-		super.createBlockStateDefinition(builder);
-	}
-
-	static final EnumSet<Direction> STICKY_X = EnumSet.of(Direction.EAST, Direction.WEST);
-	static final EnumSet<Direction> STICKY_Z = EnumSet.of(Direction.SOUTH, Direction.NORTH);
-
-	@Override
-	public EnumSet<Direction> getStickySurfaces(BlockGetter world, BlockPos pos, BlockState state) {
-		return state.getValue(BlockStateProperties.HORIZONTAL_AXIS) == Axis.X ? STICKY_X : STICKY_Z;
-	}
-
-	@Override
-	public BlockState updateShape(BlockState pState, Direction pDirection, BlockState pNeighborState,
-		LevelAccessor pLevel, BlockPos pCurrentPos, BlockPos pNeighborPos) {
-		updateWater(pLevel, pState, pCurrentPos);
-		return pState;
-	}
-
-	@Override
-	public FluidState getFluidState(BlockState pState) {
-		return fluidState(pState);
 	}
 
 	@Override
@@ -107,132 +46,13 @@ public class StandardBogeyBlock extends Block
 	}
 
 	@Override
-	public boolean allowsSingleBogeyCarriage() {
-		return true;
+	public BogeyStyle getStyle() {
+		return AllBogeyStyles.STANDARD.get();
 	}
 
 	@Override
-	public BlockState getMatchingBogey(Direction upDirection, boolean axisAlongFirst) {
-		if (upDirection != Direction.UP)
-			return null;
-		return defaultBlockState().setValue(AXIS, axisAlongFirst ? Axis.X : Axis.Z);
-	}
-
-	@Override
-	public Class<? extends BogeyRenderer> getRendererClass() {
-		return StandardBogeyRenderer.class;
-	}
-
-	@Override
-	public BogeyRenderer getRenderer() {
-		return this.renderer;
-	}
-
-	@Override
-	public boolean isTrackAxisAlongFirstCoordinate(BlockState state) {
-		return state.getValue(AXIS) == Axis.X;
-	}
-
-	@Override
-	@OnlyIn(Dist.CLIENT)
-	public void render(BlockState state, float wheelAngle, PoseStack ms, float partialTicks, MultiBufferSource buffers,
-		int light, int overlay) {
-		if (state != null) {
-			ms.translate(.5f, .5f, .5f);
-			if (state.getValue(AXIS) == Axis.X)
-				ms.mulPose(Vector3f.YP.rotationDegrees(90));
-		}
-
-		ms.translate(0, -1.5 - 1 / 128f, 0);
-
-		VertexConsumer vb = buffers.getBuffer(RenderType.cutoutMipped());
-		BlockState air = Blocks.AIR.defaultBlockState();
-
-		for (int i : Iterate.zeroAndOne)
-			CachedBufferer.block(AllBlocks.SHAFT.getDefaultState()
-				.setValue(ShaftBlock.AXIS, Axis.Z))
-				.translate(-.5f, .25f, i * -1)
-				.centre()
-				.rotateZ(wheelAngle)
-				.unCentre()
-				.light(light)
-				.renderInto(ms, vb);
-
-		if (large) {
-			renderer.render(new CompoundTag(), wheelAngle, ms, light, vb, BogeyRenderer.BogeySize.LARGE);
-		} else {
-			renderer.render(new CompoundTag(), wheelAngle, ms, light, vb, BogeyRenderer.BogeySize.SMALL);
-		}
-	}
-
-	private void renderBogey(float wheelAngle, PoseStack ms, int light, VertexConsumer vb, BlockState air) {
-		CachedBufferer.partial(AllBlockPartials.BOGEY_FRAME, air)
-			.scale(1 - 1 / 512f)
-			.light(light)
-			.renderInto(ms, vb);
-
-		for (int side : Iterate.positiveAndNegative) {
-			ms.pushPose();
-			CachedBufferer.partial(AllBlockPartials.SMALL_BOGEY_WHEELS, air)
-				.translate(0, 12 / 16f, side)
-				.rotateX(wheelAngle)
-				.light(light)
-				.renderInto(ms, vb);
-			ms.popPose();
-		}
-	}
-
-	private void renderLargeBogey(float wheelAngle, PoseStack ms, int light, VertexConsumer vb, BlockState air) {
-		for (int i : Iterate.zeroAndOne)
-			CachedBufferer.block(AllBlocks.SHAFT.getDefaultState()
-				.setValue(ShaftBlock.AXIS, Axis.X))
-				.translate(-.5f, .25f, .5f + i * -2)
-				.centre()
-				.rotateX(wheelAngle)
-				.unCentre()
-				.light(light)
-				.renderInto(ms, vb);
-
-		CachedBufferer.partial(AllBlockPartials.BOGEY_DRIVE, air)
-			.scale(1 - 1 / 512f)
-			.light(light)
-			.renderInto(ms, vb);
-		CachedBufferer.partial(AllBlockPartials.BOGEY_PISTON, air)
-			.translate(0, 0, 1 / 4f * Math.sin(AngleHelper.rad(wheelAngle)))
-			.light(light)
-			.renderInto(ms, vb);
-
-		ms.pushPose();
-		CachedBufferer.partial(AllBlockPartials.LARGE_BOGEY_WHEELS, air)
-			.translate(0, 1, 0)
-			.rotateX(wheelAngle)
-			.light(light)
-			.renderInto(ms, vb);
-		CachedBufferer.partial(AllBlockPartials.BOGEY_PIN, air)
-			.translate(0, 1, 0)
-			.rotateX(wheelAngle)
-			.translate(0, 1 / 4f, 0)
-			.rotateX(-wheelAngle)
-			.light(light)
-			.renderInto(ms, vb);
-		ms.popPose();
-	}
-
-	@Override
-	public BogeyInstance createInstance(MaterialManager materialManager, CarriageBogey bogey) {
-		if (large) {
-			return StandardBogeyInstance.drive(bogey, materialManager);
-		} else {
-			return StandardBogeyInstance.frame(bogey, materialManager);
-		}
-	}
-
-	@Override
-	public BlockState rotate(BlockState pState, Rotation pRotation) {
-		return switch (pRotation) {
-		case COUNTERCLOCKWISE_90, CLOCKWISE_90 -> pState.cycle(AXIS);
-		default -> pState;
-		};
+	public BogeyRenderer.BogeySize getSize() {
+		return large ? BogeyRenderer.BogeySize.LARGE : BogeyRenderer.BogeySize.SMALL;
 	}
 
 	@Override
@@ -250,10 +70,4 @@ public class StandardBogeyBlock extends Block
 	public BlockEntityType<? extends StandardBogeyTileEntity> getTileEntityType() {
 		return AllTileEntities.BOGEY.get();
 	}
-
-	@Override
-	public ItemRequirement getRequiredItems(BlockState state, BlockEntity te) {
-		return new ItemRequirement(ItemUseType.CONSUME, AllBlocks.RAILWAY_CASING.asStack());
-	}
-
 }
